# swagger/doc.json
{
	"openapi": "3.0.1",
	"info": { "title": "API https://tvoe-dev.ru/", "version": "v1" },
	"servers": [{ "url": "http://localhost:4000" }],
...
	"paths": {
...
		"/profile/history": {
			"get": {
				"summary": "Профиль > История просмотров",
				"tags": ["Profile"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Количествово пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей в БД"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string",
														"description": "Название фильма"
													},
													"shortDesc": {
														"type": "string",
														"description": "Краткое описание"
													},
													"trailer": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"description": "Mongoose.ObjectId"
															},
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению обложки"
															},
															"thumbnail": {
																"type": "string",
																"description": "Миниатюра (кадр из видео)"
															},
															"duration": {
																"type": "number",
																"description": "Длительность"
															}
														}
													},
													"ageLevel": {
														"type": "string",
														"description": "Возрастное ограничение"
													},
													"dateReleased": {
														"type": "Date",
														"description": "Дата релиза"
													},
													"categoryAlias": {
														"type": "string",
														"description": "Вид категориию. Может быть films, serials или collections"
													},
													"poster": {
														"type": "object",
														"properties": {
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению"
															}
														}
													},
													"rating": {
														"type": "number",
														"description": "Рейтинг от 0 до 10"
													},
													"duration": {
														"type": "string",
														"description": "Длительность фильма в секундах"
													},
													"url": {
														"type": "string",
														"description": "Путь к фильму"
													},
													"moviepagelogs": {
														"type": "object",
														"properties": {
															"_id": {
																"type": "string",
																"description": "Id таблицы moviePageLogSchema (mongoose.Schema.Types.ObjectId)"
															},
															"count": {
																"type": "number",
																"description": "Kоличество просмотров фильма"
															},
															"updatedAt": {
																"type": "array",
																"items": {
																	"type": "Date",
																	"description": "Дата обновления"
																}
															}
														}
													},
													"lastLogUpdatedAt": {
														"type": "array",
														"items": {
															"type": "Date",
															"description": "Дата последнего обновления"
														}
													}
												}
											}
										}
									}
								},
								"examples": {
									"example": {
										"value": {
											"totalSize": 2,
											"items": [
												{
													"name": "Америка: фильм",
													"shortDesc": "Вооруженный пилами Джордж Вашингтон объединяется с обожающим пиво друганом Сэмом Адамсом, чтобы приструнить британцев в ироничной комедии о Войне за независимость США.",
													"trailer": {
														"_id": "641bf64b0073f02ab65dde80",
														"src": "/videos/641bf64b0073f02ab65dde7f.mp4",
														"thumbnail": "/images/641bf64b0073f02ab65dde81.jpg"
													},
													"ageLevel": 18,
													"dateReleased": "2021-06-15",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/641f4080944e8bd192d18318.jpg"
													},
													"rating": 7,
													"duration": 5892,
													"url": "/p/amerika-film",
													"moviepagelogs": {
														"_id": "64007d7d0a569a75e14fc2a1",
														"count": 1,
														"updatedAt": [
															"2023-06-07T16:15:48.5030000Z"
														]
													},
													"lastLogUpdatedAt": ["2023-06-07T16:15:48.503Z"]
												},
												{
													"name": "Денежный выстрел: История Pornhub",
													"shortDesc": "Проникните глубже в историю самой известной платформы для взрослых Pornhub и недавний скандал вокруг неё. Детская порнография, сексуальное насилие, половая эксплуатация, торговля людьми — это лишь небольшой перечень того, что было найдено на страницах всем известного сайта",
													"trailer": {
														"_id": "6419a8a30073f02ab65d4f07",
														"src": "/videos/6419a8a30073f02ab65d4f06.mp4",
														"thumbnail": "/images/6419a8a30073f02ab65d4f08.jpg"
													},
													"ageLevel": 18,
													"dateReleased": "2023-02-17",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/63fded0598e2af31cb68b46f.jpg"
													},
													"rating": 10,
													"duration": 7063.708333,
													"url": "/p/snyat-blokirovku",
													"moviepagelogs": {
														"_id": "6419a7540073f02ab65d4e36",
														"count": 1,
														"updatedAt": [
															"2023-06-07T16:15:48.5030000Z"
														]
													},
													"lastLogUpdatedAt": ["2023-06-07T16:15:48.503Z"]
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/searchHistory": {
			"get": {
				"summary": "Админ-панель > История поиска",
				"tags": ["Admin"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Кол-во пропускаемых элементов из начала выборки"
					},
					{
						"in": "query",
						"name": "query",
						"schema": { "type": "number" },
						"description": "Для поиска документов (сущностей коллекции users), с которыми совпадают символы в имени и емейлу. Либо для полного совпадения по id (mongoose.ObjectId)  пользователя"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей в БД (документов коллекции users)"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Id (mongoose.ObjectId) поискового запроса (документа коллекции searchlogs)"
													},
													"query": {
														"type": "string",
														"description": "Запрос пользователя"
													},
													"device": {
														"type": "object",
														"properties": {
															"ip": {
																"type": "string",
																"description": "ip адресс юзера"
															},
															"os": {
																"type": "string",
																"description": "Операционная система устройства"
															},
															"isBot": {
																"type": "boolean",
																"description": "Является ли юзер ботом"
															},
															"isMobile": {
																"type": "boolean",
																"description": "Является ли дейвайс мобильным устройством"
															},
															"isDesktop": {
																"type": "boolean",
																"description": "Является ли дейвайс десктопным устройством"
															},
															"browser": {
																"type": "string",
																"description": "Название браузера"
															},
															"version": {
																"type": "string",
																"description": "Версия браузера"
															},
															"platform": {
																"type": "string",
																"description": "Платформа устройства ( ос)"
															}
														}
													},
													"updatedAt": {
														"type": "Date",
														"description": "Дата обновления"
													},
													"user": {
														"type": "object",
														"properties": {
															"_id": {
																"type": "string",
																"description": "Profile id"
															},
															"email": {
																"type": "string",
																"description": "Email user"
															},
															"avatar": {
																"type": "string",
																"description": "Path to img src profile user"
															},
															"firstname": {
																"type": "string",
																"description": "Имя этого человека"
															},
															"subscribe": {
																"type": "object",
																"description": "Информация по подписке. Если подписки нет, в этом поле null",
																"properties": {
																	"startAt": {
																		"type": "Date",
																		"description": "Дата начала подписки"
																	},
																	"finishAt": {
																		"type": "Date",
																		"description": "Дата окончания подписки"
																	},
																	"tariffId": {
																		"type": "string",
																		"description": "Id тарифа (mongoose.ObjectId)"
																	}
																}
															}
														}
													}
												}
											}
										}
									},
									"example": {
										"value": {
											"totalSize": 1539,
											"items": [
												{
													"_id": "63e7ee8b5c8c77c4bcc63003",
													"query": "Ход королевы",
													"device": {
														"ip": "84.16.240.69",
														"os": "OS X",
														"isBot": false,
														"isMobile": true,
														"isDesktop": false,
														"browser": "Safari",
														"version": "15.3",
														"platform": "iPhone"
													},
													"updatedAt": "2023-02-11T19:37:47.982Z",
													"user": {
														"_id": "63c85dad6396809788dfb949",
														"email": "khanrushlg@yandex.ru",
														"avatar": null,
														"firstname": "Рушан",
														"subscribe": {
															"startAt": "2023-02-12T04:39:28.501Z",
															"finishAt": "2023-05-13T04:39:28.501Z",
															"tariffId": "63dbe1447f457bc81bc920ca"
														}
													}
												},
												{
													"_id": "63e7ee885c8c77c4bcc62ff5",
													"query": "Х",
													"device": {
														"ip": "84.16.240.69",
														"os": "OS X",
														"isBot": false,
														"isMobile": true,
														"isDesktop": false,
														"browser": "Safari",
														"version": "15.3",
														"platform": "iPhone"
													},
													"updatedAt": "2023-02-11T19:37:44.853Z",
													"user": {
														"_id": "63c85dad6396809788dfb949",
														"email": "khanrushlg@yandex.ru",
														"avatar": null,
														"firstname": "Рушан",
														"subscribe": {
															"startAt": "2023-02-12T04:39:28.501Z",
															"finishAt": "2023-05-13T04:39:28.501Z",
															"tariffId": "63dbe1447f457bc81bc920ca"
														}
													}
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/search": {
			"get": {
				"summary": "Поиск >  Поиск фильмов, сериалов и всего их персонала сьемочной группы",
				"tags": ["Search"],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": { "type": "string" },
						"description": "Для поиска документов (сущностей таблицы movies), с которыми совпадают символы в полях name(название фильма) или origName( оригинальное название фильма), shortDesc(краткое описание), fullDesk(полное описание), countries(страна), persons(Актёры и съемочная группа)"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей(фильмов) в БД удовлетворяющих переданному query параметру"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string",
														"description": "Название фильма"
													},
													"shortDesc": {
														"type": "string",
														"description": "Краткое описание"
													},
													"trailer": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"description": "Mongoose.ObjectId"
															},
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению обложки"
															},
															"thumbnail": {
																"type": "string",
																"description": "Миниатюра (кадр из видео)"
															},
															"duration": {
																"type": "number",
																"description": "Длительность",
																"poster": {
																	"type": "object",
																	"required": false,
																	"properties": {
																		"src": {
																			"type": "string",
																			"description": "Путь к файлу изображению"
																		}
																	}
																}
															}
														}
													},
													"ageLevel": {
														"type": "string",
														"description": "Возрастное ограничение"
													},
													"dateReleased": {
														"type": "Date",
														"description": "Дата релиза"
													},
													"categoryAlias": {
														"type": "string",
														"description": "Вид категориии. Может быть films, serials или collections"
													},
													"poster": {
														"type": "object",
														"properties": {
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению"
															}
														}
													},
													"rating": {
														"type": "number",
														"description": "Рейтинг от 0 до 10"
													},
													"duration": {
														"type": "string",
														"description": "Длительность фильма в секундах"
													},
													"url": {
														"type": "string",
														"description": "Url к фильму? путь к фильму?"
													}
												}
											}
										}
									}
								},
								"examples": {
									"example": {
										"description": "Ответ в результате запроса с query параметром \"Венгрия\" (/search?query=Венгрия)",
										"value": {
											"totalSize": 3,
											"items": [
												{
													"name": "Ведьмак",
													"shortDesc": "Более чем за тысячу лет до появления Геральта из Ривии семь изгоев мира эльфов объединяют свои силы против всемогущей империи.",
													"ageLevel": 18,
													"dateReleased": "2019-12-20",
													"categoryAlias": "serials",
													"poster": {
														"src": "/images/63e00a92ccaf134edd216858.jpg"
													},
													"rating": 7,
													"duration": 56502.375,
													"url": "/p/vedmak",
													"trailer": {
														"_id": "63b4af65f5af3c4f9dca53a6",
														"src": "/videos/63b4af65f5af3c4f9dca53a5.mp4",
														"thumbnail": "/images/63b4af65f5af3c4f9dca53a7.jpg"
													}
												},
												{
													"name": "Король Англии",
													"shortDesc": "Англия, начало XV века, разгар Столетней войны. Принц Уэльский Хел ведёт распутный образ жизни со своим приятелем",
													"ageLevel": 18,
													"dateReleased": "2019-09-02",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/63c0fb1423c814afb609605a.jpg"
													},
													"rating": 8,
													"duration": 8429.083333,
													"url": "/p/korol-anglii",
													"trailer": {
														"_id": "63b4af65f5af3c4f9dca53a6",
														"src": "/videos/63b4af65f5af3c4f9dca53a5.mp4",
														"thumbnail": "/images/63b4af65f5af3c4f9dca53a7.jpg"
													}
												},
												{
													"name": "Смертельная зона",
													"shortDesc": "«Смертельная зона» — американский научно-фантастический фильм, режиссёра Микаэля Хофстрёма. Фильм вышел на Netflix 15 января 2021",
													"ageLevel": 18,
													"dateReleased": "2021-01-15",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/63c130fe23c814afb6097501.jpg"
													},
													"duration": 6952.541667,
													"url": "/p/smertelnaya-zona",
													"trailer": {
														"_id": "63b4af65f5af3c4f9dca53a6",
														"src": "/videos/63b4af65f5af3c4f9dca53a5.mp4",
														"thumbnail": "/images/63b4af65f5af3c4f9dca53a7.jpg"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}